feat: capture Claude statusline capacity (#89)
## Context

DIS-37 adds the supported Claude.ai capacity path on top of DIS-36. Claude Code supplies five-hour and seven-day subscriber rate limits to statusline commands after the first API response; Dispatch needs to capture those facts without editing operator configuration or retaining the raw statusline payload.

## What changed

- add the opt-in `dispatch-claude-statusline` stdin helper, independent of the daemon
- normalize only bounded capture metadata, a session fingerprint, model label, and decimal five-hour/seven-day windows
- atomically persist beneath `DISPATCH_HOME` with mode-600 temporary files, a cross-process lock, monotonic capture ordering, file fsync, rename, and directory fsync
- retain independently absent or currently unavailable windows with their original per-window timestamps while recording current availability separately
- merge cached windows into the provider-neutral observation by key without erasing newer facts or fabricating freshness
- document manual wrapper setup that composes with an existing Claude statusline and never edits `~/.claude/settings.json`
- keep the private OAuth usage endpoint out of the implementation

## Verification

- `just check` — 652 passed, 17 deselected; Ruff, formatting, strict mypy, package build/content checks passed
- focused capture/provider/store suite — 114 passed
- derived CLI/MCP surface suite — 58 passed
- standing review — 5/5 clean, zero open P0-P2
- targeted statusline atomicity/privacy review — 5/5 clean, zero open P0-P2
- isolated capture-to-usage smoke returned decimal windows and excluded raw session, cwd, transcript, and model identifiers
- packaged-wheel smoke confirmed the installed entrypoint is quiet and writes a normalized mode-600 snapshot

## Risks and rollout

Statusline rate limits are session-derived and can be absent before the first API response or for unsupported accounts. Dispatch reports current absence explicitly while retaining last-valid windows at their historical timestamps so normal staleness rendering remains honest. Setup is manual and opt-in; existing statuslines are not replaced.

Stack parent: #88

Linear: DIS-37

+++ src/outfitter/dispatch/core/models.py
@@ -1334,8 +1334,8 @@ class UsageWindowView(BaseModel):
     limit_id: str
     limit_name: str | None = None
     window: str
-    used_percent: int | None = None
-    remaining_percent: int | None = None
+    used_percent: float | None = None
+    remaining_percent: float | None = None
     duration_minutes: int | None = None
     resets_at: int | None = None
     reached_type: str | None = None
